fix: Sku variants sort (#2389)
## What's the purpose of this pull request?

This PR intends to fix the SKU variants sort implementation. It will
preserve Admin's variants order for cases that variants are not numbers.
(e.g. Sizes: ['S', 'M', 'L'] should be rendered as it is)

## How it works?

Example:

`{
      Size: [
        {
          ...,
          label: 'Size: L',
          value: 'L'
        },
        {
          ...,
          label: 'Size: M',
          value: 'M'
        },
        {
          ...,
          label: 'Size: S,
          value: 'S'
        },
      ]
}`

The expected result is to keep this order:

But for the following case:

`{
      Size: [
        {
          ...,
          label: 'Size: 2',
          value: '2'
        },
        {
          ...,
          label: 'Size: 3',
          value: '3'
        },
        {
          ...,
          label: 'Size: 1',
          value: '1'
        },
      ]
}`

The result should be sorted, so Sizes: [1, 2, 3]

## How to test it?

Run `yarn test` and check the results.

### Related PRs

#2374