Enable sign in Button override (#2387)
## What's the purpose of this pull request?

During the development of the B2B Self Management feature, we need to
customize and add new behaviors to the "My Account" button, making it
open a modal instead of going to the `/account` page. While most of this
development will happen in an external library, the ability to override
this button is important for us.

This PR enables just that, the ability to completely override the
default SignIn / My Account button.

- [ ] Add documentation of the override

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->

---------

Co-authored-by: Everton Ataide <everton.ataide@vtex.com>