metatag generated at (#2417)
## What's the purpose of this pull request?

create a metatag in build time in order to know when the build was
created.

## How it works?

using the faststore cli we generate a date and insert on store html

## How to test it?

finding the `data-generated-at` metatag on source code.

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `gatsby.store` AND
`nextjs.store` with this branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

## References

<!--- Spread the knowledge: is there any content you used to create this
PR that is worth sharing? --->

<!--- Extra tip: adding references to related issues or mentioning
people important to this PR may be good for the documentation and
reviewing process --->