Feat: adds new `deviceSizes` and adjust `Hero` Image (#2470)
## What's the purpose of this pull request?

This PR is part of the Performance epic and aims to add a new
`deviceSizes` with a size of `412`. This is the size used for `Moto G
Super`, the mobile device used in the PageSpees Insight (PSI).

It also changes the `Hero` Image size to fit a better config related to
the LCP metric.