Feat: Adds intersection (ViewportObserver) component (#2498)
## What's the purpose of this pull request?

This PR introduces the `ViewportObserver` component, which will be used
in future tasks as part of our performance improvement initiative.

## How it works?

We'll primarily use it on mobile devices to prevent sections outside the
viewport from being rendered initially. It checks whether a component or
section is within the viewport, based on a specified height. We're using
Moto G Power devices as the reference for this.

## How to test it?

To facilitate the test for our future tasks, I have added a `debug` mode
prop, so we can enable it for testing.
In this case, we can identify if the component is IN or OUT the
viewport. **This is only for the initial render.**

1. go to `core/src/components/product/ProductGrid/ProductGrid.tsx`, add
`debug` prop to following component:

```
  <ViewportObserver name="UIProductGrid-out-viewport" debug>
```

2. navigate to core, run `yarn dev`

3. go to the PLP, or http://localhost:3000/office?page=0

4. set the browser device to  Moto G Power
<img width="500" alt="image"
src="https://github.com/user-attachments/assets/b66c48fe-35de-4062-9265-6ca7801f935e">

5. checking via console.log when the component is IN or OUT
- Only the first row of the ProductGrid is inside the viewport, so you
should see `OUT` in the console.log. Scrolling up the next row should
appear, and you will be able to see `IN`

<img width="500" alt="image"
src="https://github.com/user-attachments/assets/e780ee08-dcaa-4899-8721-4eb12d9d8e7d">

6. checking visually when the component is IN or OUT
- Go to `Network` tab, selects `Slow 4G` as preset, refresh the page
- You will able to see a quick visual feedback (pink rectangle and blue
border) in the first ProductCard of the second row.


https://github.com/user-attachments/assets/87f902b6-2e5a-47bd-8b21-3b7897de9bdd

7. checking if the first two ProductCard image is loading=`eager` and
the rest `loading=`lazy`

<img width="500" alt="loading-images"
src="https://github.com/user-attachments/assets/76d72d87-757c-48b9-9634-c4394a93858b">


### Starters Deploy Preview


## References

https://vtex-dev.atlassian.net/browse/SFS-1507
https://github.com/vtex/faststore/pull/2404

---------

Co-authored-by: Larícia Mota <laricia.mota@vtex.com.br>