feat: Enable SWR delta and revalidate on page [slug]/p
This commit enables the experimental feature "swrDelta" in the next.config.js file, which is used for optimizing data fetching with SWR. Additionally, it adds the "revalidate" option with a value of 300 seconds (5 minutes) in the getStaticProps function of the [slug]/p.tsx file. This option controls how often the page should be revalidated and updated with new data.