Decrease FastStore CLI verbosity (#2516)
## What's the purpose of this pull request?

To decrease accidental verbosity of FastStore commands.

## How it works?

Created a logger object that's equal to `console` but overrides the
`log` function to check if the `DISCOVERY_DEBUG` flag is enabled.

## How to test it?

Install the CLI version on a test store and see how much less output is
being generated. :)

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->

---------

Co-authored-by: Emerson Laurentino <emerson@laurentino.co>