feat: next/dynamic in `CartSidebar` internal components SFS-1517 (#2553)
## What's the purpose of this pull request?

This PR is part of the performance initiative and aims to apply
`next/dynamic` to `CartSidebar` internal components.

## How to test it?

The CartSidebar should render as before. You should see the JS from
internal components loading on demand. This PR is also to prepare the
components for the time that we will render based on the
ViewportObserver/LazyLoadingSection.

### Starters Deploy Preview

- https://github.com/vtex-sites/starter.store/pull/609

Preview
- https://sfj-b6dcbf1--starter.preview.vtex.app/

### References
POC PR
- https://github.com/vtex/faststore/pull/2404