feat: split the homepage sections in own file (#2555)
Depends on 
- https://github.com/vtex/faststore/pull/2554

## What's the purpose of this pull request?

This PR is part of the performance initiative and aims to split the
homepage sections into its own file.

Let's initially separate the components of each page type into their own
files; this will help organize the import of components by page type (in
other tasks) and opens up the possibility of only importing customized
components based on the page type in the future, instead of all
components, as we currently do.

The idea is to stop using the `GlobalSections` component in the tree of
each page, but since we are not doing this for all pages yet, we’ll keep
it temporarily. We will only remove the use of this component on the
homepage (in another task). Once we’ve done this for all pages, we can
remove this component entirely, but we will keep the function
`getGlobalSectionsData`.
(packages/core/src/components/cms/GlobalSections.tsx). Global sections
will now be imported into the sections file of each page and passed
directly to `RenderSections`.


### Starters Deploy Preview

- https://github.com/vtex-sites/starter.store/pull/611

Preview
https://sfj-edd92b9--starter.preview.vtex.app/

### References
POC PR
- https://github.com/vtex/faststore/pull/2404