ci: create a job to validate the PR  title (#2580)
## What's the purpose of this pull request?

Create a workflow with a job to validate PR title. The title should
follow the [Conventional
Commits](https://www.conventionalcommits.org/en/v1.0.0/) specification.

Available types:
```
 - fix
 - feat
 - chore
 - docs
 - style
 - refactor
 - test
 - ci
```


## How it works?
![Captura de Tela 2024-12-03 às 14 35
33](https://github.com/user-attachments/assets/b9fa7903-71c3-415e-9c4b-2f4c5ddea05b)


## How to test it?

Change the title of this PR and refresh the page. When I update the
title, I need to refresh the page to see the job running again. It looks
like a visual issue.