fix: removes `useUI()` dependents sections from `LazyLoadingSection` (#2602)
## What's the purpose of this pull request?

This removes the logic that dynamically loads the sections that depend
on the `useUI()` hook: `CartSidebar` and `RegionModal`. We decided to
try to load them at another moment (//TODO: create a task for it)
instead of on click.

## How to test it?

Try to load the respective sections `CartSidebar` and `RegionModal`
using the [preview
link](https://sfj-2714681--starter.preview.vtex.app/).
You can see that both sections are being loaded in the initial loading
now. Instead of loading after clicking on the respective buttons that
trigger the sections.

### Starters Deploy Preview
- https://github.com/vtex-sites/starter.store/pull/639
- preview https://sfj-2714681--starter.preview.vtex.app/