feat: search page with SSR (#2619)
# (Don't merge!)

## What's the purpose of this pull request?

This POC is an alternative to indexing search pages using [server side
rendering](https://nextjs.org/docs/pages/building-your-application/data-fetching/get-server-side-props).
As the search page is dynamic and we need to set the title and
description based on the search term, we chose to use server side
rendering as it is the most suitable for this type of situation.

### Nextjs reference: 
**Like SSG,[ Server-Side
Rendering](https://nextjs.org/docs/basic-features/pages#server-side-rendering)
(SSR) is pre-rendered, which also makes it great for SEO. Instead of
being generated at build time, as in SSG, SSR's HTML is generated at
request time. This is great for when you have pages that are very
dynamic.**

https://nextjs.org/learn-pages-router/seo/rendering-and-ranking/rendering-strategies

## How it works?

We use the getServerSideProps function instead of getStaticProps. We
added a cache header to the response so that it is possible to cache the
return in a CDN.




## How to test it?
Click on:
https://storeframework-cm652ufll028lmgv665a6xv0g-polemunmz.b.vtex.app/s?q=headphone