feat: adjust rating styles (#2635)
## What's the purpose of this pull request?

This pull request introduces adjustments to the **Rating component** for
the new **Reviews & Ratings** section.

## How it works?

It updates the colors for both the default and actionable versions,
using `main-2` for active stars and `neutral-4` for empty stars.
Additionally, it includes the addition of a `disabled` style for the
actionable rating.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-feat-rating-field-vtex.vercel.app/review-and-ratings-playground)

## References

[JIRA TASK: SFS-2077](https://vtex-dev.atlassian.net/browse/SFS-2077)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-49445&t=97j1KIeA0brteiq8-4)

Default


![image](https://github.com/user-attachments/assets/53cff9ef-b54a-4ea6-9ea0-1d1d574ce04c)

Actionable


![image](https://github.com/user-attachments/assets/ae6395cb-2bc8-4dff-94a3-c99f0eb7c5ac)