feat: add checkbox field error variant (#2638)
## What's the purpose of this pull request?

This pull request adds styles and an error text for the CheckboxField
component. Additionally, it adjusts the styles of the CheckboxField
label. This new styling will be used in the modal for adding a review to
a product in the new Reviews and Ratings section.

## How it works?

I added the `error` and `alignment` props to the `CheckboxField`
component. The `alignment` prop is used to vertically align the
checkbox, with the default value set to `center`.

Additionally, the `label` was not previously styled, so I added styles
to it based on the Figma to ensure the label is displayed correctly.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-chore-add-checkbox-cli-vtex.vercel.app/review-and-ratings-playground)

## References

[JIRA TASK: SFS-2073](https://vtex-dev.atlassian.net/browse/SFS-2073)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-65865&t=iOanfcOwmDa4bolf-4)


![image](https://github.com/user-attachments/assets/8d1e12e2-01c2-4878-88fe-cf0b9208e1b4)