feat: add rating field (#2636)
>⚠️ THIS PR DEPENDS ON
[PR#2635](https://github.com/vtex/faststore/pull/2635) ⚠️

## What's the purpose of this pull request?

This pull request adds the RatingField component, which will initially
be used in the modal for adding reviews to a product.

## How it works?

This PR creates a new molecule component called `RatingField`. In
addition to the "input" fields for the actionable rating, it can also
receive an `id`, `label`, and `error` via props to be displayed in the
form. The component uses the `Rating` component with the `actionable`
view, passing the `onChange` handler to it.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-feat-rating-field-vtex.vercel.app//review-and-ratings-playground)

## References

[JIRA TASK: SFS-2078](https://vtex-dev.atlassian.net/browse/SFS-2078)
[JIRA TASK: SFS-2077](https://vtex-dev.atlassian.net/browse/SFS-2077)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-49445&t=iOanfcOwmDa4bolf-4)


![image](https://github.com/user-attachments/assets/56da1a26-1764-489c-90e9-2d81275715c3)


![image](https://github.com/user-attachments/assets/8351c008-3fca-4e36-ad6d-30518e005dbb)

---------

Co-authored-by: Fanny Chien <fanny.chien@vtex.com>
Co-authored-by: Fanny Chien <fannychienn@gmail.com>