feat: create textarea component (#2640)
## What's the purpose of this pull request?

This pull request adds the Textarea atom component based on the Input
component. This component will be used in the form of the modal for
adding a review to a product in the new Reviews and Ratings section.

## How does it work?

This PR introduces the `Textarea` atom component. In addition to the
standard properties, it includes a `resize` prop, which defaults to
`both`.

The styling of the `Textarea` component is based on the existing `Input`
component to maintain visual consistency.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-chore-add-textareafield-cli-vtex.vercel.app/review-and-ratings-playground)

## References

[JIRA TASK: SFS-2075](https://vtex-dev.atlassian.net/browse/SFS-2075)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-65865&t=iOanfcOwmDa4bolf-4)


![image](https://github.com/user-attachments/assets/841d3b2b-033f-49ef-a03d-1105b31d40f0)