feat: create modal footer compound component (#2672)
## What's the purpose of this pull request?

This pull request adds the ModalFooter compound component. This
component will be used inside the Modal in initially in the Modal to add
a new review in the new Reviews and Ratings section.

## How does it work?

You need to use `<ModalFooter />` inside the Modal and pass to it action
buttons. The ModalFooter is already spacing out the buttons and is
wrapping by default. You can pass `direction` prop to it, and it can be
`vertical` or `horizontal` and it is `horizontal` by default.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-feat-preview-modal-footer-vtex.vercel.app/modal-footer)

## References

[JIRA TASK: SFS-2080](https://vtex-dev.atlassian.net/browse/SFS-2080)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-66913&t=hvdVQfzTmgTLIa32-4)