feat: add tooltip component (#2644)
## What's the purpose of this pull request?

This pull request adds the Tooltip atom component. This component will
be used in the ReviewCard for displaying reviews of a product in the new
Reviews and Ratings section.

## How does it work?

This PR introduces a new molecule component, `Tooltip`, which allows
customization of its behavior and appearance. The component includes
default values for some of its props: `placement` is `"top"`,
`dismissable` is `false`, and `maxWidth` is `300`.

When `dismissable` is enabled, the tooltip uses an `IconButton` for the
dismiss action.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-chore-add-tooltip-cli-vtex.vercel.app/review-and-ratings-playground)

## References

[JIRA TASK: SFS-2070](https://vtex-dev.atlassian.net/browse/SFS-2070)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-64437&t=0rv1XYm6XsgBeDwA-4)


![image](https://github.com/user-attachments/assets/128e8aad-df02-4738-9d58-c4a2e68c14dd)

---------

Co-authored-by: Fanny Chien <fanny.chien@vtex.com>
Co-authored-by: Fanny Chien <fannychienn@gmail.com>