feat: add textareafield component (#2641)
>⚠️ THIS PR DEPENDS ON
[PR#2640](https://github.com/vtex/faststore/pull/2640) ⚠️

## What's the purpose of this pull request?

This pull request adds the TextareaField atom component based on the
InputField component. This component will be used in the form of the
modal for adding a review to a product in the new Reviews and Ratings
section.

## How does it work?

This PR introduces the `TextareaField` molecule component, which extends
the functionality of the `Textarea` component. It accepts additional
props such as `id`, `label`, `error`, and `disabled`, alongside all the
standard props of the `Textarea` component.

The `TextareaField` uses the `Textarea` component internally. One thing
to pay attention in this implementation is related to the label used in
the add new review form. In the Figma design, the label text is very
long, which could either exceed the width of the `Textarea` or break
into multiple lines, overlapping the text area.

To solve this, I added a logic using a `useEffect` hook to dynamically
calculate the width of the `Textarea` and set it as the `max-width` of
the label. Additionally, the `text-overflow: ellipsis` style is applied
to truncate the label text if it exceeds the defined width.

## How to test it?

<!--- Describe the steps with bullet points. Is there any external link
that can be used to better test it or an example? --->

### Starters Deploy Preview

<!--- Add a link to a deploy preview from `starter.store` with this
branch being used. --->

<!--- Tip: You can get an installable version of this branch from the
CodeSandbox generated when this PR is created. --->


[Preview](https://starter-git-chore-add-textareafield-cli-vtex.vercel.app/review-and-ratings-playground)

## References

[JIRA TASK: SFS-2076](https://vtex-dev.atlassian.net/browse/SFS-2076)


[Figma](https://www.figma.com/design/YXU6IX2htN2yg7udpCumZv/Reviews-%26-Ratings?node-id=131-49445&t=0rv1XYm6XsgBeDwA-4)


![image](https://github.com/user-attachments/assets/5f595a0a-1844-41ca-8113-f9076d3a14ec)

---------

Co-authored-by: Fanny Chien <fanny.chien@vtex.com>